Adapt to use latest upickle

+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/Types.scala
@@ -1,9 +1,11 @@
 package io.github.shogowada.scala.jsonrpc
 
+import io.github.shogowada.scala.jsonrpc.serializers.JSONRPCPickler.{ReadWriter => RW, macroRW, readwriter}
+
 import scala.concurrent.Future
 
 object Types {
-  type Id = Either[String, BigDecimal]
+  type Id = Either[String, Integer]
 
   type JSONSender = (String) => Future[Option[String]]
 }
